backend-python/requirements.txt
Flask==3.0.3
requests==2.32.3
google-generativeai==0.8.3
gunicorn==23.0.0
gevent==24.10.3